Import scheduler module
